Transition is working

+++ src/webgl-neuron-selector/painter/offscreen-painter/offscreen-painter.ts
@@ -1,14 +1,16 @@
+import { MorphologyData } from "./../morphology-data"
 /* eslint-disable no-bitwise */
 import {
     TgdContext,
     TgdDataset,
     TgdPainterClear,
     TgdPainterGroup,
-    TgdPainterSegments,
+    TgdPainterSegmentsMorphing,
     TgdPainterState,
     webglPresetDepth,
 } from "@tolokoban/tgd"
-import { Structure, StructureItem } from "../structure"
+
+import { StructureItem } from "../structure"
 import { MaterialIndex } from "./material-index"
 
 export class OffscreenPainter {
@@ -16,7 +18,7 @@ export class OffscreenPainter {
 
     private readonly offscreenContext: TgdContext
 
-    private _structure: Structure | undefined = undefined
+    private _data: MorphologyData | undefined = undefined
 
     private readonly group = new TgdPainterGroup()
 
@@ -33,23 +35,25 @@ export class OffscreenPainter {
         this.paint()
     }
 
-    get structure() {
-        return this._structure
+    get data() {
+        return this._data
     }
 
-    set structure(value: Structure | undefined) {
-        if (value === this._structure) return
+    set data(data: MorphologyData | undefined) {
+        if (data === this._data) return
 
-        this._structure = value
-    }
-
-    set dataset(dataset: TgdDataset) {
+        this._data = data
         this.group.delete()
+        if (!data) return
+
         const context = this.offscreenContext
-        const painter = new TgdPainterSegments(context, {
+        const datasetsPairs: [TgdDataset, TgdDataset][] = [
+            [data.dataset3D, data.datasetDendrogram],
+        ]
+        const painter = new TgdPainterSegmentsMorphing(context, {
             roundness: 3,
             minRadius: 3,
-            dataset,
+            datasetsPairs,
             material: new MaterialIndex(),
         })
         this.group.add(
@@ -62,9 +66,9 @@ export class OffscreenPainter {
     }
 
     getItemAt(xScreen: number, yScreen: number): StructureItem | null {
-        const { structure, offscreenContext: context } = this
-        if (!structure) return null
-        const data = new Uint8Array(context.width * context.height * 4)
+        const { data, offscreenContext: context } = this
+        if (!data) return null
+        const bytes = new Uint8Array(context.width * context.height * 4)
         context.gl.readPixels(
             0,
             0,
@@ -72,8 +76,9 @@ export class OffscreenPainter {
             context.height,
             context.gl.RGBA,
             context.gl.UNSIGNED_BYTE,
-            data
+            bytes
         )
+        const { structure } = data
         const [R, G, B] = context.readPixel(xScreen, yScreen)
         const value = (R + (G << 8) + (B << 16)) / 0xffffff
         const index = Math.floor((structure.length + 2) * value) - 1
